Rename app

+++ db/migrate/20170221232223_merge_data_file_into_transform.rb
@@ -1,7 +1,7 @@
 class MergeDataFileIntoTransform < ActiveRecord::Migration
   def up
     # Nuke these b/c only the Demo Workflow exists at this time, and it's easier to just re-seed
-    WorkflowSeeder.demo_workflow.destroy!
+    WorkflowSeeder.demo_workflow.destroy! if WorkflowSeeder.demo_workflow_exists?
 
     add_column :transforms, :s3_region_name, :string
     add_column :transforms, :s3_bucket_name, :string
